refactor: review comments
