+++ src/schema/users.ts
@@ -32,7 +32,10 @@ import { IResolvers } from '@graphql-tools/utils';
 import { FileUpload } from 'graphql-upload/GraphQLUpload.js';
 import { AuthContext, BaseContext, Context } from '../Context';
 import { traceResolverObject } from './trace';
-import { queryPaginatedByDate } from '../common/datePageGenerator';
+import {
+  GQLDatePageGeneratorConfig,
+  queryPaginatedByDate,
+} from '../common/datePageGenerator';
 import {
   getInviteLink,
   getShortUrl,
@@ -71,9 +74,9 @@ import { FastifyBaseLogger } from 'fastify';
 import { cachePrefillMarketingCta } from '../common/redisCache';
 import { cio } from '../cio';
 import {
-  UserIntegrationType,
   UserIntegration,
   UserIntegrationSlack,
+  UserIntegrationType,
 } from '../entity/UserIntegration';
 
 export interface GQLUpdateUserInput {
@@ -560,6 +563,20 @@ export const typeDefs = /* GraphQL */ `
     name: String!
   }
 
+  type UserIntegrationEdge {
+    node: UserIntegration!
+
+    """
+    Used in \`before\` and \`after\` args
+    """
+    cursor: String!
+  }
+
+  type UserIntegrationConnection {
+    pageInfo: PageInfo!
+    edges: [UserIntegrationEdge!]!
+  }
+
   extend type Query {
     """
     Get user based on logged in session
@@ -697,7 +714,7 @@ export const typeDefs = /* GraphQL */ `
     """
     Get integrations for the user
     """
-    userIntegrations: [UserIntegration]! @auth
+    userIntegrations: UserIntegrationConnection @auth
   }
 
   extend type Mutation {
@@ -956,18 +973,6 @@ const getUserStreakQuery = async (
         id: id,
       }),
   }));
-};
-
-const getUserIntegrationName = (userIntegration: UserIntegration) => {
-  switch (userIntegration.type) {
-    case UserIntegrationType.Slack: {
-      const slackIntegration = userIntegration as UserIntegrationSlack;
-
-      return slackIntegration.meta.teamName;
-    }
-    default:
-      return userIntegration.type;
-  }
 };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -1331,22 +1336,31 @@ export const resolvers: IResolvers<any, BaseContext> = {
     },
     userIntegrations: async (
       _,
-      __,
-      ctx: Context,
-    ): Promise<GQLUserIntegration[]> => {
-      const userIntegrations = await ctx.con
-        .getRepository(UserIntegration)
-        .find({ where: { userId: ctx.userId } });
-
-      return userIntegrations.map((item) => {
-        return {
-          id: item.id,
-          type: item.type,
-          createdAt: item.createdAt,
-          updatedAt: item.updatedAt,
-          name: getUserIntegrationName(item),
-        };
-      });
+      args: ConnectionArguments,
+      ctx: AuthContext,
+      info,
+    ): Promise<Connection<GQLUserIntegration>> => {
+      return queryPaginatedByDate(
+        ctx,
+        info,
+        args,
+        { key: 'createdAt' } as GQLDatePageGeneratorConfig<
+          GQLUserIntegration,
+          'createdAt'
+        >,
+        {
+          queryBuilder: (builder) => {
+            builder.queryBuilder = builder.queryBuilder.andWhere(
+              `${builder.alias}."userId" = :integrationUserId`,
+              {
+                integrationUserId: ctx.userId,
+              },
+            );
+            return builder;
+          },
+          orderByKey: 'DESC',
+        },
+      );
     },
   }),
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -1662,5 +1676,18 @@ export const resolvers: IResolvers<any, BaseContext> = {
   }),
   User: {
     permalink: getUserPermalink,
+  },
+  UserIntegration: {
+    name: (userIntegration: UserIntegration) => {
+      switch (userIntegration.type) {
+        case UserIntegrationType.Slack: {
+          const slackIntegration = userIntegration as UserIntegrationSlack;
+
+          return slackIntegration.meta.teamName;
+        }
+        default:
+          return userIntegration.type;
+      }
+    },
   },
 };
